front: move cutSpaceTimeChart function in its own file
